+++ run/src/main/java/com/guide/run/event/controller/EventSearchController.java
@@ -7,7 +7,8 @@ import com.guide.run.event.service.EventSearchService;
 import com.guide.run.event.service.EventService;
 import com.guide.run.global.exception.user.resource.NotExistUserException;
 import com.guide.run.global.jwt.JwtProvider;
-import com.guide.run.user.repository.UserRepository;
+
+import com.guide.run.user.repository.user.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
