+++ run/src/main/java/com/guide/run/event/entity/dto/response/get/Count.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 @AllArgsConstructor
 @Builder
 public class Count {
-    private int count;
+    private long count;
 }
