+++ run/src/test/java/com/guide/run/user/service/UserServiceTest.java
@@ -19,13 +19,13 @@ class UserServiceTest {
     UserRepository userRepository;
 
     @BeforeEach
-    public void init(){
+    public void init() {
         userRepository.deleteAll();
     }
 
     @DisplayName("로그인 시 기존 회원 응답")
     @Test
-    void existUserLoginResponse(){
+    void existUserLoginResponse() {
         User user = User.builder()
                 .userId(userService.getUUID())
                 .userId("kakao_1")
@@ -38,7 +38,7 @@ class UserServiceTest {
 
     @DisplayName("로그인 시 가입 대기중 회원 응답")
     @Test
-    void waitUserLoginResponse(){
+    void waitUserLoginResponse() {
         User user = User.builder()
                 .userId(userService.getUUID())
                 .userId("kakao_1")
@@ -48,3 +48,4 @@ class UserServiceTest {
         boolean userStatus = userService.getUserStatus("kakao_1");
         Assertions.assertThat(userStatus).isEqualTo(Role.ROLE_WAIT.getValue());
     }
+}
